requirements.txt
azure-ai-contentsafety==1.0.0
azure-core==1.30.0
certifi==2024.2.2
charset-normalizer==3.3.2
flake8==7.0.0
idna==3.6
isodate==0.6.1
mccabe==0.7.0
pycodestyle==2.11.1
pyflakes==3.2.0
python-dotenv==1.0.1
requests==2.31.0
six==1.16.0
typing_extensions==4.9.0
urllib3==2.2.1
